extract radarPreferenceFragment to other file
