+++ app/src/main/java/com/denysnovoa/nzbmanager/settings/screen/view/screen/SettingsActivity.kt
@@ -1,7 +1,9 @@
 package com.denysnovoa.nzbmanager.settings.screen.view.screen
 
+import android.os.Build
 import android.preference.PreferenceActivity
-import org.jetbrains.anko.startActivity
+import android.view.MenuItem
+import com.denysnovoa.nzbmanager.R
 
 /**
  * A [PreferenceActivity] that presents a set of application settings. On
@@ -34,7 +36,7 @@ class SettingsActivity : AppCompatPreferenceActivity() {
         return isXLargeTablet(this)
     }
 
-    override fun onOptionsItemSelected(item: android.view.MenuItem) = when (item.itemId) {
+    override fun onOptionsItemSelected(item: MenuItem) = when (item.itemId) {
         android.R.id.home -> {
             onBackPressed()
             true
@@ -43,46 +45,16 @@ class SettingsActivity : AppCompatPreferenceActivity() {
     }
 
     /** {@inheritDoc}  */
-    @android.annotation.TargetApi(android.os.Build.VERSION_CODES.HONEYCOMB)
+    @android.annotation.TargetApi(Build.VERSION_CODES.HONEYCOMB)
     override fun onBuildHeaders(target: List<android.preference.PreferenceActivity.Header>) {
-        loadHeadersFromResource(com.denysnovoa.nzbmanager.R.xml.pref_headers, target)
+        loadHeadersFromResource(R.xml.pref_headers, target)
     }
 
     override fun isValidFragment(fragmentName: String): Boolean {
         return android.preference.PreferenceFragment::class.java.name == fragmentName
-                || com.denysnovoa.nzbmanager.settings.screen.view.screen.SettingsActivity.RadarPreferenceFragment::class.java.name == fragmentName
+                || RadarPreferenceFragment::class.java.name == fragmentName
     }
 
-    @android.annotation.TargetApi(android.os.Build.VERSION_CODES.HONEYCOMB)
-    class RadarPreferenceFragment : android.preference.PreferenceFragment() {
-        override fun onCreate(savedInstanceState: android.os.Bundle?) {
-            super.onCreate(savedInstanceState)
-            addPreferencesFromResource(com.denysnovoa.nzbmanager.R.xml.pref_radarr)
-            setHasOptionsMenu(true)
-
-            //bindPreferenceSummaryToValue(findPreference("pref_radarr_hots_address"))
-            //bindPreferenceSummaryToValue(findPreference("pref_radarr_hots_port"))
-            //bindPreferenceSummaryToValue(findPreference("pref_radarr_api_key"))
-        }
-
-        override fun onOptionsItemSelected(item: android.view.MenuItem) = when (item.itemId) {
-            android.R.id.home -> {
-                startActivity<SettingsActivity>(); true
-            }
-            else -> super.onOptionsItemSelected(item)
-
-        }
-
-
-//                : Boolean {
-//            val id = item.itemId
-//            if (id == android.R.id.home) {
-//                startActivity(android.content.Intent(activity, SettingsActivity::class.java))
-//                return true
-//            }
-//            return super.onOptionsItemSelected(item)
-//        }
-    }
 
     /**
      * Helper method to determine if the device has an extra-large screen. For
