+++ app/src/main/res/xml/pref_headers.xml
@@ -1,7 +1,7 @@
 <preference-headers xmlns:android="http://schemas.android.com/apk/res/android">
   <header
       android:summary="@string/settings_summary_radarr"
-      android:fragment="com.denysnovoa.nzbmanager.settings.screen.view.screen.SettingsActivity$RadarPreferenceFragment"
+      android:fragment="com.denysnovoa.nzbmanager.settings.screen.view.screen.RadarPreferenceFragment"
       android:title="@string/pref_header_radarr"
       android:icon="@drawable/ic_info_black_24dp"
       />
